+++ NEWS.md
@@ -1,7 +1,6 @@
 # mvMAPIT (development version)
 
-# mvMAPIT 2.0.0-alpha prerelease
-Interface changed.
+* Added `Dockerfile` and vignette on mvMAPIT in Docker.
 * `MvMAPIT` was renamed to `mvmapit`.
 * `simulate_phenotypes` was renamed to `simulate_traits`.
 * Dependencies were cleaned up, e.g. R requirement 2.10 -> 3.5, old package dependencies removed.